+++ commonmark-ext-footnotes/src/main/java/org/commonmark/ext/footnotes/internal/FootnoteBlockParser.java
@@ -39,6 +39,10 @@ public class FootnoteBlockParser extends AbstractBlockParser {
         if (parserState.getIndent() >= 4) {
             // It looks like content needs to be indented by 4 so that it's part of a footnote (instead of starting a new block).
             return BlockContinue.atColumn(4);
+        } else if (parserState.isBlank()) {
+            // A blank line doesn't finish a footnote yet. If there's another line with indent >= 4 after it,
+            // that should result in another paragraph in this footnote definition.
+            return BlockContinue.atIndex(parserState.getIndex());
         } else {
             // We're not continuing to give other block parsers a chance to interrupt this definition.
             // But if no other block parser applied (including another FootnotesBlockParser), we will