+++ commonmark-ext-footnotes/src/main/java/org/commonmark/ext/footnotes/internal/FootnoteMarkdownNodeRenderer.java
@@ -55,9 +55,7 @@ public class FootnoteMarkdownNodeRenderer implements NodeRenderer {
         writer.raw("]: ");
 
         writer.pushPrefix("    ");
-        writer.pushTight(true);
         renderChildren(def);
-        writer.popTight();
         writer.popPrefix();
     }